Move getFirstFrameOfVideo to it's own file.
