+++ assets/src/edit-story/utils/useUploadVideoFrame.js
@@ -8,6 +8,7 @@ import { useCallback } from '@wordpress/element';
 import { useAPI } from '../app/api';
 import { useStory } from '../app/story';
 import { useConfig } from '../app/config';
+import getFirstFrameOfVideo from './getFirstFrameOfVideo';
 
 function useUploadVideoFrame( { videoId, src, id } ) {
 	const { actions: { uploadMedia, saveMedia } } = useAPI();
@@ -17,40 +18,9 @@ function useUploadVideoFrame( { videoId, src, id } ) {
 		( properties ) => updateElementById( { elementId: id, properties } ),
 		[ id, updateElementById ] );
 
-	/**
-	 * Returns an image of the first frame of a given video.
-	 *
-	 * @param {string} src Video src URL.
-	 * @return {Promise<string>} The extracted image in base64-encoded format.
-	 */
-	const getFirstFrameOfVideo = useCallback( () => {
-		const video = document.createElement( 'video' );
-		video.muted = true;
-		video.crossOrigin = 'anonymous';
-		video.preload = 'metadata';
-		video.currentTime = 0.5; // Needed to seek forward.
-
-		return new Promise( ( resolve, reject ) => {
-			video.addEventListener( 'error', reject );
-
-			video.addEventListener( 'canplay', () => {
-				const canvas = document.createElement( 'canvas' );
-				canvas.width = video.videoWidth;
-				canvas.height = video.videoHeight;
-
-				const ctx = canvas.getContext( '2d' );
-				ctx.drawImage( video, 0, 0, canvas.width, canvas.height );
-
-				canvas.toBlob( resolve, 'image/jpeg' );
-			} );
-
-			video.src = src;
-		} );
-	}, [ src ] );
-
 	const processData = async () => {
 		try {
-			const obj = await getFirstFrameOfVideo();
+			const obj = await getFirstFrameOfVideo( src );
 			const { id: featuredMedia, source_url: featuredMediaSrc } = await uploadMedia( obj );
 			await saveMedia( featuredMedia, {
 				meta: {
